Minion stat rebalancing

+++ Assets/Scripts/GameManager.cs
@@ -14,7 +14,7 @@ public class GameManager : Singleton<GameManager>
     //Constant for the default next wave starting score bonus/remaining wave time. 
     //Note: Prefer to keep at 15 seconds from now on
     private const int Countdown = 15;
-    public int DifficultyMultipler;
+    public float DifficultyMultipler;
     //Remaining units until the next block of the wave progress bar
     private decimal RemainingUnits { get; set; } = 1.83m;
     //Remaining time left in wave
